Add #:strict? keyword argument
When `#:changed` and `#:removed` contain assets that are not in the
graph in a call to `compile!`, the associated `unlike-compiler%` instance would
always raise an exception. This makes users have to filter out lists in advance.

Add a flag to allow users to disable the exception and cause `compile!`
to no-op on invalid assets instead.